[REFACTORING] Split IMAPServerPlainAuthDisabledTest
